+++ src/main/scala/org/justcards/server/user_manager/User.scala
@@ -1,9 +1,8 @@
 package org.justcards.server.user_manager
 
-import akka.actor.{Actor, ActorRef, Props}
+import akka.actor.{ActorRef, Props}
 import akka.io.Tcp._
 import org.justcards.commons._
-import org.justcards.commons.AppMessage._
 import org.justcards.commons.actor_connection.{ActorWithConnection, ActorWithTcp, Outer}
 import org.justcards.server
 import org.justcards.server.user_manager.UserManagerMessage.{LogOutAndExitFromLobby, UserLogout}
@@ -68,6 +67,8 @@ object User {
 
   private[this] class UserActorWithTcp(private val userRef: ActorRef, private val userManager: ActorRef)
     extends BasicUserActor(userRef, userManager) with ActorWithTcp {
+
+    import org.justcards.commons.actor_connection.ActorWithTcp._
 
     override def errorBehaviour(username: String): Receive = {
       case CommandFailed(w: Write) =>
